Attempted to create calendar functionality.
Tried to add code to create functionality with calendar view and created bills.

Added line 29 to Bill.cs and lines 12-30 for this purpose.

Added lines 96-116 to try and create this functionality, but attempt was unsuccessful. Lines were commented out to save the attempt and work on later.

Additionally, some comment lines were added around these changes to explain additions or provide context.